==> codes/5- Descriptive table.Rmd <==
---
title: "5- Descriptive table"
author: "Tesneem Fnais"
date: "2025-08-11"
output: html_document
---

```{r setup, message=FALSE, warning=FALSE}
library(gtsummary)
library(dplyr)
library(forcats)

train_data <- readRDS("data/train_data.rds")
test_data  <- readRDS("data/test_data.rds")

full_data <- dplyr::bind_rows(train_data, test_data)

# --- Step 3: Recode Variables ---
data_clean <- full_data %>%
  mutate(
    heartoa = factor(heartoa, levels = c("no_OA", "OA"), labels = c("No OA", "OA")),
    sex = factor(sex, labels = c("Male", "Female")),
    hepaa = factor(hepaa, levels = 0:3, labels = c("No pain", "Mild", "Moderate", "Severe")),
    hehelf = factor(hehelf, levels = 1:5, labels = c("Excellent", "Very good", "Good", "Fair", "Poor")),
    
    across(
      c(hepawba, hepawhi, hepawkn, hepawfe, hepawot,
        mmpain, hemobwa, hemobsi, hemobch, hemobcs,
        hemobcl, hemobst, hemobre, hemobpu, hemobpi,
        headldr, headlba, headlbe, headlsh, headlwc),
      ~ factor(.x, levels = c(0, 1), labels = c("No", "Yes"))
    )
  )

# --- Step 4: Create Stratified Table ---
table1 <- data_clean %>%
  select(
    age, sex, hehelf, grip_mean,
    hepaa, mmpain, hepawba, hepawhi, hepawkn, hepawfe, hepawot,
    hemobst, hemobch, hemobcs, hemobcl, hemobwa, hemobsi, hemobpu, hemobre, hemobpi,
    headldr, headlba, headlbe, headlsh, headlwc, 
    heartoa
  ) %>%
  tbl_summary(
    by = heartoa,
    statistic = list(all_continuous() ~ "{mean} ({sd})"),
    digits = all_continuous() ~ 2,
    missing = "no",
    label = list(
      age ~ "Age (years)",
      sex ~ "Sex",
      hehelf ~ "Self-Rated General Health",
      grip_mean ~ "Grip Strength (kg)",
      
      hepaa ~ "Pain Severity (Most of the time)",
      mmpain ~ "Pain While Walking",
      hepawba ~ "Back Pain",
      hepawhi ~ "Hip Pain",
      hepawkn ~ "Knee Pain",
      hepawfe ~ "Foot Pain",
      hepawot ~ "Pain Elsewhere",
      
      hemobst ~ "Difficulty Stooping/Kneeling",
      hemobch ~ "Difficulty Getting Up from Chair",
      hemobcs ~ "Difficulty Climbing Several Flights",
      hemobcl ~ "Difficulty Climbing One Flight",
      hemobwa ~ "Difficulty Walking 100 Yards",
      hemobsi ~ "Difficulty Sitting 2 Hours",
      hemobpu ~ "Difficulty Pushing Objects",
      hemobre ~ "Difficulty Reaching Overhead",
      hemobpi ~ "Difficulty Lifting 10 lbs",
      
      headldr ~ "Difficulty Dressing",
      headlba ~ "Difficulty Bathing",
      headlbe ~ "Difficulty Getting In/Out of Bed",
      headlsh ~ "Difficulty Shopping For Groceries", 
      headlwc ~ "Difficulty Using The Toilet"
    )
  ) %>%
  add_p() %>%
  modify_header(label ~ "**Variable**") %>%
  bold_labels()

# --- Step 5: View Table ---
table1
```

```{r}
# --- Step 6: Load Flextable Package ---
library(flextable)

# --- Step 7: Save Table to Word Document ---
as_flex_table(table1) %>%
  flextable::save_as_docx(path = "Table1_OA_corrected.docx")
```
